backend/requirements.txt
django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
psycopg2-binary==2.9.9
channels==4.1.0
channels-redis==4.2.0
celery==5.4.0
redis==5.0.4
django-redis==5.4.0
